chore(size): export <Index /> from its own widget
This makes /dom, /react-native "pure modules", which _might have advantages in webpack 4 cc @thelarkinn

We'll also need `side-effects: false` in our package.json, see [example](https://github.com/webpack/webpack/tree/next/examples/side-effects), but I'm not sure if we're _actually_ side-effects RN, could you confirm Sean?
