+++ packages/react-instantsearch/native.js
@@ -1,13 +1,2 @@
-import algoliasearch from 'algoliasearch/reactnative';
-import createInstantSearch from './src/core/createInstantSearch';
-import createIndex from './src/core/createIndex';
-import { View } from 'react-native';
-const InstantSearch = createInstantSearch(algoliasearch, {
-  Root: View,
-});
-export { InstantSearch };
-const Index = createIndex({
-  Root: View,
-});
-export { Index };
+export { default as Index } from './src/widgets/Index.native.js';
 export { default as Configure } from './src/widgets/Configure.js';
